+++ src/data/users/usersDataBase.json
@@ -158,5 +158,35 @@
     "password_repeat": "$2a$10$gdLKB4IlCADlGwpQJznAmuuwXRgKuc8A4fBJsY59HzVbGI36bP7ru",
     "category": "user",
     "avatar": "avatar-1696191910301.jpg"
+  },
+  {
+    "id": "927ffd53-12a1-4555-93ca-3781cd5d6e88",
+    "name": "pablo",
+    "last_name": "gomez middleware",
+    "email": "otromail@mail.com",
+    "password": "$2a$10$iucZpj17z25nivD2qaoyFuITLfflC284v.QxnwtlzB/hvWuq7VTnu",
+    "password_repeat": "$2a$10$R1Y4eP4WxTGPngug8WWiY.Jj.mhXCucI61wvuh7y7xI/vQTPXjOBe",
+    "category": "user",
+    "avatar": "users_default.png"
+  },
+  {
+    "id": "8dfecf8d-29c4-48f3-ae4e-9edc63cc87ec",
+    "name": "pablo",
+    "last_name": "gomez middleware",
+    "email": "otromail@mail.com",
+    "password": "$2a$10$pjCtXfaNSXwg8P/OTv9axuXJLgSNjasRHXNP2sDKVewQsPaTJJ6Ca",
+    "password_repeat": "$2a$10$IlbShDMUsX1oFcIQ2A4ST.NZWdaEAtPesebzQrYH.g.AeINRUOrYW",
+    "category": "user",
+    "avatar": "avatar-1696200295274.jpg"
+  },
+  {
+    "id": "1f880b89-32c4-42c2-96e8-ed946cb0af61",
+    "name": "final prueba",
+    "last_name": "con middleware",
+    "email": "mail@mail.com",
+    "password": "$2a$10$i/v3KCumttRdeNCBy0xu6OaUOsVzpKFbiOkXcl2qQD8QpFM9eGaEm",
+    "password_repeat": "$2a$10$tsJFqLzVIAgL9xLKJEgqs.yFiG7fExehzR35DMxZv2O7rdKfzNZFW",
+    "category": "user",
+    "avatar": "avatar-1696200556291.jpg"
   }
 ]
