+++ src/routes/user-router.js
@@ -1,41 +1,9 @@
 // ************ Require's ************
 const { Router } = require("express");
 const router = Router();
+const validationsRegister = require("../validations/register-validations");
+const validateFormRegister = require("../middlewares/validate-form-register");
 const path = require("path");
-
-const { body } = require("express-validator");
-
-const validations = [
-  body("name").notEmpty().withMessage("Tienes que escribir un nombre"),
-  body("last_name").notEmpty().withMessage("Tienes que escribir un apellido"),
-  body("email")
-    .notEmpty()
-    .withMessage("Tienes que escribir un correo eléctrónico")
-    .bail()
-    .isEmail()
-    .withMessage("Debes ingresar un formato de correo válido"),
-  body("password").notEmpty().withMessage("Tienes que escribir una contraseña"),
-  body("password_repeat")
-    .notEmpty()
-    .withMessage("Tienes que repetir la contraseña"),
-  body("avatar").custom((value, { req }) => {
-    let file = req.file;
-    let acceptedExtensions = [".jpg", ".png", ".gif"];
-
-    if (!file) {
-      return true;
-      /*throw new Error("Tienes que subir una imagen");*/
-    } else {
-      let fileExtension = path.extname(file.originalname);
-      if (!acceptedExtensions.includes(fileExtension)) {
-        throw new Error(
-          `Las extensiones aceptadas son ${acceptedExtensions.join(", ")}`
-        );
-      }
-    }
-    return true;
-  }),
-];
 
 const multer = require("multer");
 
@@ -62,7 +30,8 @@ router.get("/register/", userController.register);
 router.post(
   "/register/",
   upload.single("avatar"),
-  validations,
+  validationsRegister,
+  validateFormRegister,
   userController.createUser
 );
